requirements.txt
PyQt5==5.15.11
pyperclip==1.9.0
pycryptodome==3.20.0